+++ android/koin-android/src/main/java/org/koin/androidx/viewmodel/scope/ScopeExt.kt
@@ -17,12 +17,10 @@ package org.koin.androidx.viewmodel.scope
 
 import android.os.Bundle
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelStore
 import androidx.savedstate.SavedStateRegistryOwner
-import org.koin.androidx.viewmodel.ViewModelOwnerDefinition
-import org.koin.androidx.viewmodel.ViewModelParameter
-import org.koin.androidx.viewmodel.createViewModelProvider
-import org.koin.androidx.viewmodel.resolveInstance
+import org.koin.androidx.viewmodel.*
 import org.koin.core.parameter.ParametersDefinition
 import org.koin.core.qualifier.Qualifier
 import org.koin.core.scope.Scope
@@ -33,12 +31,8 @@ import kotlin.reflect.KClass
  *
  * @author Arnaud Giuliani
  */
-typealias SavedStateRegistryOwnerDefinition = () -> SavedStateRegistryOwner
-typealias ViewModelStoreDefinition = () -> ViewModelStore
-
 fun emptyState(): BundleDefinition = { Bundle() }
 typealias BundleDefinition = () -> Bundle
-
 
 inline fun <reified T : ViewModel> Scope.viewModel(
         qualifier: Qualifier? = null,
@@ -82,6 +76,6 @@ fun <T : ViewModel> Scope.getViewModel(
 }
 
 fun <T : ViewModel> Scope.getViewModel(viewModelParameters: ViewModelParameter<T>): T {
-    val viewModelProvider = createViewModelProvider(viewModelParameters)
+    val viewModelProvider = ViewModelProvider(viewModelParameters.viewModelStore, pickFactory(viewModelParameters))
     return viewModelProvider.resolveInstance(viewModelParameters)
 }